Fixed tooltip bugs
